# cmp: more accurate handling of large and small numbers